requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.1
pydantic[email]>=1.8.2
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
python-multipart>=0.0.5
aioredis>=2.0.0
python-dotenv>=0.19.0
numpy>=1.21.0
APScheduler>=3.10.1

# 测试依赖
pytest>=8.0.0
pytest-asyncio>=0.23.2
httpx>=0.26.0
h11>=0.14.0
